+++ trunk/utils/xhead/xhead.cpp
@@ -19,7 +19,7 @@ int xrecv(int sock, xhead_t* xhead, const uint32_t buffsize,
 	timeout.tv_sec  = timeo_ms/1000;
 	timeout.tv_usec = 1000*(timeo_ms%1000);
 	setsockopt(sock, SOL_SOCKET, SO_RCVTIMEO, &timeout, sizeof(timeval));
-	int ret = recv(sock, (char*)xhead, sizeof(xhead_t), MSG_WAITALL);
+	int ret = (int)recv(sock, (char*)xhead, sizeof(xhead_t), MSG_WAITALL);
 	if (ret != sizeof(xhead_t))
 	{
         ALARM("read xhead_t error. ret[%d] errno[%d] sizeof_xhead_t[%u] to_s[%u] to_us[%u] msg[%m]",
@@ -35,7 +35,7 @@ int xrecv(int sock, xhead_t* xhead, const uint32_t buffsize,
         ALARM("buffer too short. buffsize[%u] detail_len[%u]", buffsize, xhead->detail_len);
 		return -3;
 	}
-	ret = recv(sock, (char*)(&xhead[1]), xhead->detail_len, MSG_WAITALL);
+	ret = (int)recv(sock, (char*)(&xhead[1]), xhead->detail_len, MSG_WAITALL);
 	if(ret != (int)xhead->detail_len)
 	{
         ALARM("read detail error. logid[%u] ret[%d] errno[%d] detail_len[%u] to_s[%u] to_us[%u] msg[%m]",
@@ -51,7 +51,7 @@ int SendNByte(const int fd, const char* buff, const uint32_t size, uint32_t flag
 	uint32_t offset = size;
 	while (left > 0)
 	{
-		int len = send(fd, buff+offset-left, left, flag);
+		int len = (int)send(fd, buff+offset-left, left, flag);
 		if ((len == -1 || len == 0 ) && errno != EAGAIN)
 		{
 			return -1;
@@ -67,7 +67,7 @@ int xsend(int sock, const xhead_t* xhead, const uint32_t timeo_ms)
 	{
 		return -1;
 	}
-	int size = sizeof(xhead_t) + xhead->detail_len;
+	int size = int(sizeof(xhead_t) + xhead->detail_len);
 	struct timeval timeout = {0, 0};
 	timeout.tv_sec  = timeo_ms/1000;
 	timeout.tv_usec = 1000*(timeo_ms%1000);
